+++ src/integrations/index.ts
@@ -1,2 +1,6 @@
 export { creduentVerifyToolDefinition } from "./vercel.js";
 export { verifyAgentNode } from "./langgraph.js";
+export { verifyCreduentAgent } from "./langchain.js";
+export { verifyLlamaIndexAgent } from "./llamaindex.js";
+
+